% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gesis.R
\name{browse_codebook}
\alias{browse_codebook}
\title{Browse dataset codebook}
\usage{
browse_codebook(doi, browseURL = TRUE, ...)
}
\arguments{
\item{doi}{The dataset's DOI.}

\item{browseURL}{If FALSE, returns codebook's URL instead of opening in
browser.}

\item{...}{Additional arguments passed to \code{browseURL}.}
}
\description{
Open a dataset's codebook in browser
}
\examples{
\dontrun{browse_codebook("5928")}
}

