% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis.R
\name{download_codebook}
\alias{download_codebook}
\title{Download the codebook for a Gesis data set}
\usage{
download_codebook(doi, path = ".", quiet = FALSE)
}
\arguments{
\item{doi}{The unique identifier(s) for the data set(s)}

\item{path}{Directory to which to download the file}

\item{quiet}{Whether to output download message.}
}
\value{
Nothing
}
\description{
Download the codebook for a Gesis data set
}
\examples{
download_codebook(doi = "0078")
}
