% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis.R
\name{gesis_download}
\alias{gesis_download}
\title{Streamlined downloads from GESIS}
\usage{
gesis_download(doi, filetype = "dta", user = getOption("gesis_user"),
  pass = getOption("gesis_pass"), purpose = 1,
  download_dir = "gesis_data", msg = TRUE)
}
\arguments{
\item{doi}{The unique identifier (or optionally a vector of these identifiers)
for the dataset(s) to be downloaded (see details).}

\item{filetype}{The filetype to be downloaded (usually only "dta" or "spss" are available).}

\item{user, pass}{Your GESIS username and password (see details)}

\item{purpose}{The purpose for which you are downloading the data set (see details).}

\item{download_dir}{The directory (relative to your working directory) to
which files from GESIS will be downloaded.}

\item{msg}{If TRUE, outputs a message showing which data set is being downloaded.}
}
\value{
The function returns downloaded files.
}
\description{
\code{gesis_download} provides a programmatic and reproducible means to download datasets from GESIS
}
\details{
Datasets reposited with GESIS are uniquely identified with a
  numberic identifier called a "DOI". This identifier appears both in the URL
  for a dataset's website, and on the website itself.

  In addition to accepting the terms of use, you need to input a purpose for
  downloading a data set. The options are as follows:

1. for scientific research (incl. PhD)
2. for research with commercial mandate
3. for teaching as lecturer
4. for my academic studies
5. for my final exam (e.g. bachelor or master)
6. for professional training and qualification

 To avoid requiring others to edit your scripts to insert their own email and
 password, the default is set to fetch this information from the user's
 .Rprofile.  Before running \code{gesis_download}, then, you should be sure to
 add these options to your .Rprofile substituting your info for the example below:

 \code{
  options("gesis_user" = "juanita-herrara@uppermidwest.edu",
         "gesis_pass" = "password123!")
 }
}
\examples{
\dontrun{
 gesis_download(doi = c(5900, 5928))
}

}

