% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis.R
\name{get_datasets}
\alias{get_datasets}
\title{Get a dataframe of all individual data sets within a group of studies}
\usage{
get_datasets(group_no)
}
\arguments{
\item{group_no}{The group number (usually obtained from get_study_groups())}
}
\value{
A dataframe
}
\description{
Get a dataframe of all individual data sets within a group of studies
}
\examples{
# Get DOIs and titles for all Eurobarometer studies
eurobars <- get_datasets("0008")
head(eurobars)
}

