% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis.R
\name{get_study_groups}
\alias{get_study_groups}
\title{Get a dataframe with all available groups of studies}
\usage{
get_study_groups()
}
\value{
A dataframe
}
\description{
Get a dataframe with all available groups of studies
}
\examples{
\dontrun{
groups <- get_study_groups()
head(groups)
}
}
