% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis.R
\name{login}
\alias{login}
\title{Log in to the Gesis website}
\usage{
login(username = "", password = "")
}
\arguments{
\item{username}{Your Gesis username}

\item{password}{Your Gesis password}
}
\value{
A session object
}
\description{
Log in to the Gesis website
}
\details{
The username and password can also be stored as enviroenment variables
"GESIS_USER" and "GESIS_PASS" so as not to store these in plaintext in a
script.
}
\examples{
\dontrun{s <- login("my_gesis_username", "my_gesis_password")}
}

