% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gesis.R
\name{login_gesis}
\alias{login_gesis}
\title{Log in to GESIS}
\usage{
login_gesis(remDr, user = getOption("gesis_user"),
  pass = getOption("gesis_pass"))
}
\arguments{
\item{remDr}{The remote driver object created with \code{setup_gesis}.}

\item{user,pass}{Your GESIS user name and password.}
}
\value{
Nothing.
}
\description{
Create connection with GESIS and log in.
}
\examples{
\dontrun{
gesis_remDr <- setup_gesis(download_dir = "downloads")
login_gesis(gesis_remDr, user = "myusername", pass = "mypassword")
download_dataset(gesis_remDr, doi = 5928)
}
}

